requirements.txt
crypto==1.4.1
Django==1.10.2
ecdsa==0.13
Fabric3==1.12.post1
Naked==0.1.31
paramiko==1.17.2
pycrypto==2.6.1
PyYAML==3.12
requests==2.11.1
shellescape==3.4.1
six==1.10.0
pip freeze
